// src/modules/auth/constants/role.enum.ts
export enum Role {
  ADMIN = "admin",
  SUPERADMIN = "superadmin",
  EMPLOYEE = "employee",
  SYSTEMADMIN = "systemadmin",
  ORGANIZATION = "organization",
}

export enum OrganizationRole {
  ADMIN = "admin",
  SUPERADMIN = "superadmin",
}

export enum SystemAdmin {
  SYSTEMADMIN = "systemadmin",
}

export enum Organization {
  ORGANIZATION = "organization",
}